moved AFFECTED constant to AffectedTestsDetector class
