+++ strategies/affected/src/test/java/org/arquillian/smart/testing/strategies/affected/AffectedConfigurationTest.java
@@ -7,7 +7,7 @@ import org.arquillian.smart.testing.configuration.ConfigurationLoader;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.arquillian.smart.testing.strategies.affected.AffectedConfiguration.AFFECTED;
+import static org.arquillian.smart.testing.strategies.affected.AffectedTestsDetector.AFFECTED;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class AffectedConfigurationTest {
