+++ strategies/affected/src/test/java/org/arquillian/smart/testing/strategies/affected/AffectedTestsDetectorTest.java
@@ -24,7 +24,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import static org.arquillian.smart.testing.strategies.affected.AffectedConfiguration.AFFECTED;
+import static org.arquillian.smart.testing.strategies.affected.AffectedTestsDetector.AFFECTED;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
